# Use Account.local.sum(statuses_count) instead of Status.local.count