@ykzts 2017-09-18
Could you paste benchmark?