@akihikodaki 2017-09-18
Sure.

Before:
```
> EXPLAIN ANALYZE SELECT COUNT(*) FROM statuses WHERE statuses.local OR statuses.uri IS NULL;
QUERY PLAN
Aggregate  (cost=2097206.31..2097206.32 rows=1 width=0) (actual time=211399.367..211399.367 rows=1 loops=1)
  ->  Seq Scan on statuses  (cost=0.00..2079733.00 rows=6989323 width=0) (actual time=0.015..209753.089 rows=7333522 loops=1)
        Filter: (local OR (uri IS NULL))
        Rows Removed by Filter: 30663909
Planning time: 1.157 ms
Execution time: 211401.419 ms
```

After:
```
> EXPLAIN ANALYZE SELECT SUM(statuses_count) FROM accounts WHERE domain IS NULL;
QUERY PLAN
Aggregate  (cost=59544.74..59544.75 rows=1 width=4) (actual time=5376.827..5376.828 rows=1 loops=1)
  ->  Seq Scan on accounts  (cost=0.00..58832.52 rows=284886 width=4) (actual time=0.526..5255.530 rows=285078 loops=1)
        Filter: (domain IS NULL)
        Rows Removed by Filter: 176648
Planning time: 1.995 ms
Execution time: 5376.909 ms
```

So it is approximately 40 times faster.