@akihikodaki 2017-09-18
This removes the first two queries shown in #4888.